+++ bin/pubspec_extract.dart
@@ -12,14 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// @dart=1.9
-
 import 'dart:io';
 
 import 'package:args/args.dart';
 import 'package:logging/logging.dart';
 import 'package:path/path.dart' as p;
-
 import 'package:pubspec_extract/pubspec_extract.dart';
 
 int main(List<String> arguments) {