+++ lib/pubspec_extract.dart
@@ -12,6 +12,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export 'src/builder.dart';
 export 'src/generator.dart';
 export 'src/pubspec.dart';